requirements.txt
Django==1.9.6

django-appconf==1.0.1
djangorestframework==3.3
